Added basic map
I added a basic map activity.  Currently, it takes in a dummy restaurant
object and creates a map with a marker to that location.  I also shows
your current location and asks for permissions if it is not already
there.

The map activity launches by hitting the red button.  In the end
product, the button should instead launch an activity with restaurant
info and then that activity should be able to launch the maps one.